only evaluate print-modifiers once
